main/requirements.txt
fastapi
uvicorn[standard]
pydantic[dotenv]
websockets
python-multipart
faster-whisper
vosk
langchain
faiss-cpu
sentence-transformers
piper-tts
detoxify
langdetect
transformers
accelerate
torch
torchvision
torchaudio
gpt4all
openai-whisper
playsound
SpeechRecognition
PyAudio
soundfile
sounddevice
ffmpeg
flask
flask-cors
python-dotenv
